+++ js/const.js
@@ -1,0 +1,22 @@
+const MESSAGE_EXAMPLE=`Всё отлично!
+В целом всё неплохо. Но не всё.
+Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.
+Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.
+Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.
+Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!`;
+
+const ARRAY_COMMENTS = MESSAGE_EXAMPLE.split('\n');
+
+const SURNAMES = [
+  'Марья',
+  'Верон',
+  'Мирабелла',
+  'Вальц',
+  'Онопко',
+  'Топольницкая',
+  'Нионго',
+  'Ирвинг',
+];
+
+
+export {ARRAY_COMMENTS, SURNAMES};
