+++ js/util.js
@@ -1,0 +1,13 @@
+function getRandomNumber(min, max) {
+  return max >= 0 && min >= 0
+    ? Math.round(Math.random() * (max - min) + min)
+    : null;
+}
+
+
+function controlLenghtString(str, maxlength = 140) {
+  return str.length <= maxlength;
+}
+
+
+export {getRandomNumber, controlLenghtString};
